isolate Android projects too
